LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/html/semantics/forms/the-input-element/type-change-state-expected.txt


PASS change state from hidden to text
PASS change state from hidden to search
PASS change state from hidden to tel
PASS change state from hidden to url
PASS change state from hidden to email
PASS change state from hidden to password
FAIL change state from hidden to datetime-local assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from hidden to date assert_implements: Support for input type date is required for this test. undefined
FAIL change state from hidden to month assert_implements: Support for input type month is required for this test. undefined
FAIL change state from hidden to week assert_implements: Support for input type week is required for this test. undefined
FAIL change state from hidden to time assert_implements: Support for input type time is required for this test. undefined
PASS change state from hidden to number
PASS change state from hidden to range
FAIL change state from hidden to color assert_implements: Support for input type color is required for this test. undefined
PASS change state from hidden to checkbox
PASS change state from hidden to radio
PASS change state from hidden to file
PASS change state from hidden to submit
PASS change state from hidden to image
PASS change state from hidden to reset
PASS change state from hidden to button
PASS change state from text to hidden
PASS change state from text to search
PASS change state from text to tel
PASS change state from text to url
PASS change state from text to email
PASS change state from text to password
FAIL change state from text to datetime-local assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from text to date assert_implements: Support for input type date is required for this test. undefined
FAIL change state from text to month assert_implements: Support for input type month is required for this test. undefined
FAIL change state from text to week assert_implements: Support for input type week is required for this test. undefined
FAIL change state from text to time assert_implements: Support for input type time is required for this test. undefined
PASS change state from text to number
PASS change state from text to range
FAIL change state from text to color assert_implements: Support for input type color is required for this test. undefined
PASS change state from text to checkbox
PASS change state from text to radio
PASS change state from text to file
PASS change state from text to submit
PASS change state from text to image
PASS change state from text to reset
PASS change state from text to button
PASS change state from search to hidden
PASS change state from search to text
PASS change state from search to tel
PASS change state from search to url
PASS change state from search to email
PASS change state from search to password
FAIL change state from search to datetime-local assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from search to date assert_implements: Support for input type date is required for this test. undefined
FAIL change state from search to month assert_implements: Support for input type month is required for this test. undefined
FAIL change state from search to week assert_implements: Support for input type week is required for this test. undefined
FAIL change state from search to time assert_implements: Support for input type time is required for this test. undefined
PASS change state from search to number
PASS change state from search to range
FAIL change state from search to color assert_implements: Support for input type color is required for this test. undefined
PASS change state from search to checkbox
PASS change state from search to radio
PASS change state from search to file
PASS change state from search to submit
PASS change state from search to image
PASS change state from search to reset
PASS change state from search to button
PASS change state from tel to hidden
PASS change state from tel to text
PASS change state from tel to search
PASS change state from tel to url
PASS change state from tel to email
PASS change state from tel to password
FAIL change state from tel to datetime-local assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from tel to date assert_implements: Support for input type date is required for this test. undefined
FAIL change state from tel to month assert_implements: Support for input type month is required for this test. undefined
FAIL change state from tel to week assert_implements: Support for input type week is required for this test. undefined
FAIL change state from tel to time assert_implements: Support for input type time is required for this test. undefined
PASS change state from tel to number
PASS change state from tel to range
FAIL change state from tel to color assert_implements: Support for input type color is required for this test. undefined
PASS change state from tel to checkbox
PASS change state from tel to radio
PASS change state from tel to file
PASS change state from tel to submit
PASS change state from tel to image
PASS change state from tel to reset
PASS change state from tel to button
PASS change state from url to hidden
PASS change state from url to text
PASS change state from url to search
PASS change state from url to tel
PASS change state from url to email
PASS change state from url to password
FAIL change state from url to datetime-local assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from url to date assert_implements: Support for input type date is required for this test. undefined
FAIL change state from url to month assert_implements: Support for input type month is required for this test. undefined
FAIL change state from url to week assert_implements: Support for input type week is required for this test. undefined
FAIL change state from url to time assert_implements: Support for input type time is required for this test. undefined
PASS change state from url to number
PASS change state from url to range
FAIL change state from url to color assert_implements: Support for input type color is required for this test. undefined
PASS change state from url to checkbox
PASS change state from url to radio
PASS change state from url to file
PASS change state from url to submit
PASS change state from url to image
PASS change state from url to reset
PASS change state from url to button
PASS change state from email to hidden
PASS change state from email to text
PASS change state from email to search
PASS change state from email to tel
PASS change state from email to url
PASS change state from email to password
FAIL change state from email to datetime-local assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from email to date assert_implements: Support for input type date is required for this test. undefined
FAIL change state from email to month assert_implements: Support for input type month is required for this test. undefined
FAIL change state from email to week assert_implements: Support for input type week is required for this test. undefined
FAIL change state from email to time assert_implements: Support for input type time is required for this test. undefined
PASS change state from email to number
PASS change state from email to range
FAIL change state from email to color assert_implements: Support for input type color is required for this test. undefined
PASS change state from email to checkbox
PASS change state from email to radio
PASS change state from email to file
PASS change state from email to submit
PASS change state from email to image
PASS change state from email to reset
PASS change state from email to button
PASS change state from password to hidden
PASS change state from password to text
PASS change state from password to search
PASS change state from password to tel
PASS change state from password to url
PASS change state from password to email
FAIL change state from password to datetime-local assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from password to date assert_implements: Support for input type date is required for this test. undefined
FAIL change state from password to month assert_implements: Support for input type month is required for this test. undefined
FAIL change state from password to week assert_implements: Support for input type week is required for this test. undefined
FAIL change state from password to time assert_implements: Support for input type time is required for this test. undefined
PASS change state from password to number
PASS change state from password to range
FAIL change state from password to color assert_implements: Support for input type color is required for this test. undefined
PASS change state from password to checkbox
PASS change state from password to radio
PASS change state from password to file
PASS change state from password to submit
PASS change state from password to image
PASS change state from password to reset
PASS change state from password to button
FAIL change state from datetime-local to hidden assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to text assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to search assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to tel assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to url assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to email assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to password assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to date assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to month assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to week assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to time assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to number assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to range assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to color assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to checkbox assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to radio assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to file assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to submit assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to image assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to reset assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from datetime-local to button assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from date to hidden assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to text assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to search assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to tel assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to url assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to email assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to password assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to datetime-local assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to month assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to week assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to time assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to number assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to range assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to color assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to checkbox assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to radio assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to file assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to submit assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to image assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to reset assert_implements: Support for input type date is required for this test. undefined
FAIL change state from date to button assert_implements: Support for input type date is required for this test. undefined
FAIL change state from month to hidden assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to text assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to search assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to tel assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to url assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to email assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to password assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to datetime-local assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to date assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to week assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to time assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to number assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to range assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to color assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to checkbox assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to radio assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to file assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to submit assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to image assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to reset assert_implements: Support for input type month is required for this test. undefined
FAIL change state from month to button assert_implements: Support for input type month is required for this test. undefined
FAIL change state from week to hidden assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to text assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to search assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to tel assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to url assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to email assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to password assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to datetime-local assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to date assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to month assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to time assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to number assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to range assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to color assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to checkbox assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to radio assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to file assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to submit assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to image assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to reset assert_implements: Support for input type week is required for this test. undefined
FAIL change state from week to button assert_implements: Support for input type week is required for this test. undefined
FAIL change state from time to hidden assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to text assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to search assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to tel assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to url assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to email assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to password assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to datetime-local assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to date assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to month assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to week assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to number assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to range assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to color assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to checkbox assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to radio assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to file assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to submit assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to image assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to reset assert_implements: Support for input type time is required for this test. undefined
FAIL change state from time to button assert_implements: Support for input type time is required for this test. undefined
PASS change state from number to hidden
PASS change state from number to text
PASS change state from number to search
PASS change state from number to tel
PASS change state from number to url
PASS change state from number to email
PASS change state from number to password
FAIL change state from number to datetime-local assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from number to date assert_implements: Support for input type date is required for this test. undefined
FAIL change state from number to month assert_implements: Support for input type month is required for this test. undefined
FAIL change state from number to week assert_implements: Support for input type week is required for this test. undefined
FAIL change state from number to time assert_implements: Support for input type time is required for this test. undefined
PASS change state from number to range
FAIL change state from number to color assert_implements: Support for input type color is required for this test. undefined
FAIL change state from number to checkbox assert_equals: input.value should be 'on' after change of state expected "on" but got ""
FAIL change state from number to radio assert_equals: input.value should be 'on' after change of state expected "on" but got ""
PASS change state from number to file
PASS change state from number to submit
PASS change state from number to image
PASS change state from number to reset
PASS change state from number to button
PASS change state from range to hidden
PASS change state from range to text
PASS change state from range to search
PASS change state from range to tel
PASS change state from range to url
PASS change state from range to email
PASS change state from range to password
FAIL change state from range to datetime-local assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from range to date assert_implements: Support for input type date is required for this test. undefined
FAIL change state from range to month assert_implements: Support for input type month is required for this test. undefined
FAIL change state from range to week assert_implements: Support for input type week is required for this test. undefined
FAIL change state from range to time assert_implements: Support for input type time is required for this test. undefined
PASS change state from range to number
FAIL change state from range to color assert_implements: Support for input type color is required for this test. undefined
PASS change state from range to checkbox
PASS change state from range to radio
PASS change state from range to file
PASS change state from range to submit
PASS change state from range to image
PASS change state from range to reset
PASS change state from range to button
FAIL change state from color to hidden assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to text assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to search assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to tel assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to url assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to email assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to password assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to datetime-local assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to date assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to month assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to week assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to time assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to number assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to range assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to checkbox assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to radio assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to file assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to submit assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to image assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to reset assert_implements: Support for input type color is required for this test. undefined
FAIL change state from color to button assert_implements: Support for input type color is required for this test. undefined
PASS change state from checkbox to hidden
PASS change state from checkbox to text
PASS change state from checkbox to search
PASS change state from checkbox to tel
PASS change state from checkbox to url
PASS change state from checkbox to email
PASS change state from checkbox to password
FAIL change state from checkbox to datetime-local assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from checkbox to date assert_implements: Support for input type date is required for this test. undefined
FAIL change state from checkbox to month assert_implements: Support for input type month is required for this test. undefined
FAIL change state from checkbox to week assert_implements: Support for input type week is required for this test. undefined
FAIL change state from checkbox to time assert_implements: Support for input type time is required for this test. undefined
PASS change state from checkbox to number
PASS change state from checkbox to range
FAIL change state from checkbox to color assert_implements: Support for input type color is required for this test. undefined
PASS change state from checkbox to radio
PASS change state from checkbox to file
PASS change state from checkbox to submit
PASS change state from checkbox to image
PASS change state from checkbox to reset
PASS change state from checkbox to button
PASS change state from radio to hidden
PASS change state from radio to text
PASS change state from radio to search
PASS change state from radio to tel
PASS change state from radio to url
PASS change state from radio to email
PASS change state from radio to password
FAIL change state from radio to datetime-local assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from radio to date assert_implements: Support for input type date is required for this test. undefined
FAIL change state from radio to month assert_implements: Support for input type month is required for this test. undefined
FAIL change state from radio to week assert_implements: Support for input type week is required for this test. undefined
FAIL change state from radio to time assert_implements: Support for input type time is required for this test. undefined
PASS change state from radio to number
PASS change state from radio to range
FAIL change state from radio to color assert_implements: Support for input type color is required for this test. undefined
PASS change state from radio to checkbox
PASS change state from radio to file
PASS change state from radio to submit
PASS change state from radio to image
PASS change state from radio to reset
PASS change state from radio to button
PASS change state from file to hidden
PASS change state from file to text
PASS change state from file to search
PASS change state from file to tel
PASS change state from file to url
PASS change state from file to email
PASS change state from file to password
FAIL change state from file to datetime-local assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from file to date assert_implements: Support for input type date is required for this test. undefined
FAIL change state from file to month assert_implements: Support for input type month is required for this test. undefined
FAIL change state from file to week assert_implements: Support for input type week is required for this test. undefined
FAIL change state from file to time assert_implements: Support for input type time is required for this test. undefined
PASS change state from file to number
PASS change state from file to range
FAIL change state from file to color assert_implements: Support for input type color is required for this test. undefined
PASS change state from file to checkbox
PASS change state from file to radio
PASS change state from file to submit
PASS change state from file to image
PASS change state from file to reset
PASS change state from file to button
PASS change state from submit to hidden
PASS change state from submit to text
PASS change state from submit to search
PASS change state from submit to tel
PASS change state from submit to url
PASS change state from submit to email
PASS change state from submit to password
FAIL change state from submit to datetime-local assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from submit to date assert_implements: Support for input type date is required for this test. undefined
FAIL change state from submit to month assert_implements: Support for input type month is required for this test. undefined
FAIL change state from submit to week assert_implements: Support for input type week is required for this test. undefined
FAIL change state from submit to time assert_implements: Support for input type time is required for this test. undefined
PASS change state from submit to number
PASS change state from submit to range
FAIL change state from submit to color assert_implements: Support for input type color is required for this test. undefined
PASS change state from submit to checkbox
PASS change state from submit to radio
PASS change state from submit to file
PASS change state from submit to image
PASS change state from submit to reset
PASS change state from submit to button
PASS change state from image to hidden
PASS change state from image to text
PASS change state from image to search
PASS change state from image to tel
PASS change state from image to url
PASS change state from image to email
PASS change state from image to password
FAIL change state from image to datetime-local assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from image to date assert_implements: Support for input type date is required for this test. undefined
FAIL change state from image to month assert_implements: Support for input type month is required for this test. undefined
FAIL change state from image to week assert_implements: Support for input type week is required for this test. undefined
FAIL change state from image to time assert_implements: Support for input type time is required for this test. undefined
PASS change state from image to number
PASS change state from image to range
FAIL change state from image to color assert_implements: Support for input type color is required for this test. undefined
PASS change state from image to checkbox
PASS change state from image to radio
PASS change state from image to file
PASS change state from image to submit
PASS change state from image to reset
PASS change state from image to button
PASS change state from reset to hidden
PASS change state from reset to text
PASS change state from reset to search
PASS change state from reset to tel
PASS change state from reset to url
PASS change state from reset to email
PASS change state from reset to password
FAIL change state from reset to datetime-local assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from reset to date assert_implements: Support for input type date is required for this test. undefined
FAIL change state from reset to month assert_implements: Support for input type month is required for this test. undefined
FAIL change state from reset to week assert_implements: Support for input type week is required for this test. undefined
FAIL change state from reset to time assert_implements: Support for input type time is required for this test. undefined
PASS change state from reset to number
PASS change state from reset to range
FAIL change state from reset to color assert_implements: Support for input type color is required for this test. undefined
PASS change state from reset to checkbox
PASS change state from reset to radio
PASS change state from reset to file
PASS change state from reset to submit
PASS change state from reset to image
PASS change state from reset to button
PASS change state from button to hidden
PASS change state from button to text
PASS change state from button to search
PASS change state from button to tel
PASS change state from button to url
PASS change state from button to email
PASS change state from button to password
FAIL change state from button to datetime-local assert_implements: Support for input type datetime-local is required for this test. undefined
FAIL change state from button to date assert_implements: Support for input type date is required for this test. undefined
FAIL change state from button to month assert_implements: Support for input type month is required for this test. undefined
FAIL change state from button to week assert_implements: Support for input type week is required for this test. undefined
FAIL change state from button to time assert_implements: Support for input type time is required for this test. undefined
PASS change state from button to number
PASS change state from button to range
FAIL change state from button to color assert_implements: Support for input type color is required for this test. undefined
PASS change state from button to checkbox
PASS change state from button to radio
PASS change state from button to file
PASS change state from button to submit
PASS change state from button to image
PASS change state from button to reset
